Add space after non-final argument completions
Also tag them with `is_completion=True` to prevent early command execution.

Use `has_placeholder_item=True` (the default) to allow commands to be executed before all arguments have been entered. Thre are pathological argument configurations `like VarArgs(Choice(...))` that may not behave intuitively with this change, but fixing that is left for a later exercise when it's needed.